Move website backend into adaptive-oracle for refactor to backend
