+++ README.md
@@ -9,6 +9,5 @@ The basic strategy is to use statistical modelling to predict confirmation times
 
 Note:  you need to create a folder 'json' in the parent directory and then a new json file will be written to the folder each block containing the gas price predictions and the prediction table.  The gas price in ethgasAPI is currently in 10gwei units, so divide them by 10 to get in gwei.
 
-usage:  `python3 ethgasstation.py`
-
+usage:  `python3 gasStationFull.py`
 requirements: `pip3 install -r requirements.txt`
