+++ modelparams.py
@@ -11,11 +11,20 @@ You should have about >20k transactions in the database to fit the model accurat
 #interaction term with highgas offered and hashpower. not currently using
 #highgas offefred coefficient
 
-INTERCEPT = 7.5980 
-HPA_COEF = -0.0811
+INTERCEPT = 7.5375
+HPA_COEF = -0.0801
 TXATABOVE_COEF = 0.0003 
 INTERACT_COEF = 0 
-HIGHGAS_COEF = 0.4085
+HIGHGAS_COEF = 0.3532
+
+
+#a second model when data from past 5 min is available
+INT2 = 5.5751
+HPA2 = -0.0454
+TXATAB2 = 0
+HIGHGAS2 = 0.6658
+S5MAGO = 0.0176
+
 #high gas offered is defined based as a percentage of the gas limit
 #highgas2 is the only one that matters right now
